refactor: move ErrorView folder src/view/ProfileView/ to src/view/
