+++ src/view/ErrorView/data.ts
@@ -1,0 +1,4 @@
+/* eslint-disable import/prefer-default-export */
+export enum DataAttrs {
+  ERROR_TEXT = 'data-error-text',
+}
